fix(ci): resolve queue and MCP check regressions
