+++ src/auto-reply/reply/agent-runner-run.ts
@@ -1,10 +1,4 @@
-import { expectDefined } from "@openclaw/normalization-core";
-import { normalizeOptionalString } from "@openclaw/normalization-core/string-coerce";
 import { resolveDefaultAgentId } from "../../agents/agent-scope-config.js";
-import {
-  formatEmbeddedAgentQueueFailureSummary,
-  queueEmbeddedAgentMessageWithOutcomeAsync,
-} from "../../agents/embedded-agent-runner/runs.js";
 import { settleProgressVisibilityCallbackResult } from "../../channels/progress-visibility.js";
 import { hasRestartRecoverySourceClaim } from "../../config/sessions/restart-recovery-state.js";
 import { loadSessionEntry, updateSessionEntry } from "../../config/sessions/session-accessor.js";
@@ -33,7 +27,7 @@ import {
   isAudioPayload,
 } from "./agent-runner-helpers.js";
 import { resetReplyRunSession } from "./agent-runner-session-reset.js";
-import { finalizeAcceptedSteer } from "./agent-runner-steer-adoption.js";
+import { runActiveReplySteer } from "./agent-runner-steer-adoption.js";
 import { resolveQueuedReplyExecutionConfig } from "./agent-runner-utils.js";
 import { createAudioAsVoiceBuffer, createBlockReplyPipeline } from "./block-reply-pipeline.js";
 import { resolveEffectiveBlockStreamingConfig } from "./block-streaming.js";
@@ -46,18 +40,12 @@ import { createFollowupRunner } from "./followup-runner.js";
 import { REPLY_RUN_STILL_SHUTTING_DOWN_TEXT } from "./get-reply-run-queue.js";
 import { resolveOriginMessageProvider } from "./origin-routing.js";
 import { resolveActiveRunQueueAction } from "./queue-policy.js";
-import {
-  admitFollowupRunLifecycle,
-  enqueueFollowupRun,
-  parkSteerCandidate,
-  resolveFollowupAbortSignal,
-  scheduleFollowupDrain,
-} from "./queue.js";
+import { enqueueFollowupRun, scheduleFollowupDrain } from "./queue.js";
 import { REPLY_ADMISSION_TICKET } from "./reply-admission-ticket.js";
 import { createReplyMediaContext } from "./reply-media-paths.js";
 import * as replyRunState from "./reply-operation-run-state.js";
 import { type ReplyOperation, replyRunRegistry } from "./reply-run-registry.js";
-import { bindReplyOperationTyping, refreshReplyOperationTyping } from "./reply-run-typing.js";
+import { bindReplyOperationTyping } from "./reply-run-typing.js";
 import { createReplyToModeFilterForChannel, resolveReplyToMode } from "./reply-threading.js";
 import { admitReplyTurn, resolveReplyTurnKind } from "./reply-turn-admission.js";
 import {
@@ -65,7 +53,7 @@ import {
   retireTerminalRestartRecoverySourceClaim,
 } from "./restart-recovery-claim.js";
 import { resolveRoutedDeliveryThreadId } from "./routed-delivery-thread.js";
-import { buildChannelSourceTurnId, readChannelSourceTurnId } from "./source-turn-id.js";
+import { readChannelSourceTurnId } from "./source-turn-id.js";
 import { createTypingSignaler } from "./typing-mode.js";
 export async function runReplyAgent(
   params: RunReplyAgentParams,
@@ -235,155 +223,24 @@ export async function runReplyAgent(
   });
 
   if (effectiveShouldSteer && isActive && opts?.messageInjectionAttempted !== true) {
-    // Steer against the operation that owns THIS session's run slot. A native
-    // command continuation whose slot adoption was skipped (#104844) still
-    // carries a source-keyed reservation; steering by its stale sessionId
-    // would miss the live target run.
-    const registeredReplyOperation = sessionKey ? replyRunRegistry.get(sessionKey) : undefined;
-    const activeReplyOperation =
-      providedReplyOperation?.key === sessionKey
-        ? providedReplyOperation
-        : (registeredReplyOperation ?? providedReplyOperation);
-    const steerSessionId = activeReplyOperation?.sessionId ?? followupRun.run.sessionId;
     replyRunState.bindQueueDispositionToRunState(followupRun, replyOperationRunState);
-    const parked = parkSteerCandidate(queueKey, followupRun, resolvedQueue, queuedRunFollowupTurn);
-    if (!parked) {
-      releaseAdmissionTicket();
-      typing.cleanup();
-      return undefined;
-    }
-    const scheduleParkedFallback = () => {
-      const owner = replyRunRegistry.get(queueKey);
-      if (owner) {
-        scheduleFollowupDrainAfterReplyOperationClear({
-          operation: owner,
-          queueKey,
-          runFollowup: queuedRunFollowupTurn,
-        });
-      } else {
-        scheduleFollowupDrain(queueKey, queuedRunFollowupTurn);
-      }
-    };
-    scheduleParkedFallback();
-    releaseAdmissionTicket();
-    try {
-      const admission = await parked.admit();
-      if (admission === "cancelled") {
-        parked.consume();
-        typing.cleanup();
-        return undefined;
-      }
-      if (admission === "fallback") {
-        parked.fallback();
-        if (replyOperationRunState) {
-          replyOperationRunState.admission = { status: "accepted", mode: "followup" };
-        }
-        await touchActiveSessionEntry();
-        typing.cleanup();
-        return undefined;
-      }
-      // Channel dispatch normally stamps the route-scoped source id. Internal
-      // callers can derive the same per-message identity from the prepared turn.
-      const steerRunId = expectDefined(
-        restartRecoverySourceTurnId ??
-          buildChannelSourceTurnId({
-            provider:
-              followupRun.originatingChannel ??
-              followupRun.run.messageProvider ??
-              sessionCtx.Provider,
-            accountId:
-              followupRun.originatingAccountId ??
-              followupRun.run.agentAccountId ??
-              sessionCtx.AccountId,
-            conversationId:
-              followupRun.originatingTo ??
-              followupRun.originatingChatId ??
-              sessionKey ??
-              followupRun.run.sessionKey,
-            messageId: followupRun.messageId ?? sessionCtx.MessageSidFull ?? sessionCtx.MessageSid,
-          }) ??
-          normalizeOptionalString(opts?.runId),
-        "steered turn id",
-      );
-      const steerOutcome = await queueEmbeddedAgentMessageWithOutcomeAsync(
-        steerSessionId,
-        followupRun.prompt,
-        {
-          steeringMode: "all",
-          isInboundUserMessage: true,
-          ...(followupRun.images?.length ? { images: followupRun.images } : {}),
-          ...(followupRun.imageOrder?.length ? { imageOrder: followupRun.imageOrder } : {}),
-          ...(followupRun.media?.length ? { media: followupRun.media } : {}),
-          waitForTranscriptCommit: true,
-          queueIdentity: steerRunId,
-          abortSignal: resolveFollowupAbortSignal(followupRun),
-          onQueueAccepted: parked.accepted,
-          ...(resolvedQueue.debounceMs !== undefined
-            ? { debounceMs: resolvedQueue.debounceMs }
-            : {}),
-          ...(followupRun.run.sourceReplyDeliveryMode
-            ? { sourceReplyDeliveryMode: followupRun.run.sourceReplyDeliveryMode }
-            : {}),
-          taskSuggestionDeliveryMode: followupRun.run.taskSuggestionDeliveryMode,
-          ...(followupRun.userTurnTranscriptRecorder
-            ? { userTurnTranscriptRecorder: followupRun.userTurnTranscriptRecorder }
-            : {}),
-        },
-      );
-      if (!steerOutcome.queued) {
-        parked.fallback();
-        if (replyOperationRunState) {
-          replyOperationRunState.admission = { status: "accepted", mode: "followup" };
-        }
-        const summary = formatEmbeddedAgentQueueFailureSummary(steerOutcome);
-        logVerbose(
-          `queue: active session ${steerSessionId} rejected steering injection: ${summary}`,
-        );
-        await touchActiveSessionEntry();
-        typing.cleanup();
-        return undefined;
-      }
-      const adoptionDisposition = await finalizeAcceptedSteer({
-        activeReplyOperation,
-        abortKey: sessionKey ?? queueKey,
-        cleanupTyping: () => typing.cleanup(),
-        errorMessage: steerOutcome.errorMessage,
-        onAdopted: () => admitFollowupRunLifecycle(followupRun),
-        replyOperationRunState,
-        steerSessionId,
-        transcriptCommit: steerOutcome.transcriptCommit,
-      });
-      parked.consume();
-      if (adoptionDisposition === "stop") {
-        return undefined;
-      }
-      if (followupRun.currentInboundAudio === true) {
-        activeReplyOperation?.markAcceptedSteeredInboundAudio();
-      }
-      if (activeReplyOperation) {
-        await refreshReplyOperationTyping(activeReplyOperation, {
-          startIfIdle: typingSignals.shouldStartImmediately,
-        });
-      }
-      await touchActiveSessionEntry();
-      typing.cleanup();
-      return undefined;
-    } catch (error) {
-      if (resolveFollowupAbortSignal(followupRun)?.aborted) {
-        parked.consume();
-      } else {
-        parked.fallback();
-      }
-      throw error;
-    } finally {
-      if (followupRun.steerPending) {
-        if (resolveFollowupAbortSignal(followupRun)?.aborted) {
-          parked.consume();
-        } else {
-          parked.fallback();
-        }
-      }
-    }
+    await runActiveReplySteer({
+      followupRun,
+      opts,
+      providedReplyOperation,
+      queueKey,
+      releaseAdmissionTicket,
+      replyOperationRunState,
+      resolvedQueue,
+      restartRecoverySourceTurnId,
+      runFollowup: queuedRunFollowupTurn,
+      sessionCtx,
+      sessionKey,
+      touchActiveSessionEntry,
+      typing,
+      typingSignals,
+    });
+    return undefined;
   }
 
   const activeRunQueueAction = resolveActiveRunQueueAction({
